ml/requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
psycopg2-binary==2.9.9
joblib==1.4.2
python-dotenv==1.0.1
numpy==1.26.4
